requirements.txt
python-socketio[aiohttp]
aiohttp
pymongo
dnspython
PyQt5
